Fixed an overflow which occurs in hashtable size computations. Fixes #4678.
No review.

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@25112 5e8d7ff9-d8ef-0310-90f0-a4852d11357a

+++ src/library/scala/collection/mutable/FlatHashTable.scala
@@ -321,9 +321,9 @@ private[collection] object FlatHashTable {
   
   /** The initial size of the hash table.
    */
-  private[collection] def initialSize: Int = 16  
+  private[collection] def initialSize: Int = 16
   
-  private[collection] def sizeForThreshold(size: Int, _loadFactor: Int) = size * loadFactorDenum / _loadFactor
+  private[collection] def sizeForThreshold(size: Int, _loadFactor: Int) = (size.toLong * loadFactorDenum / _loadFactor).toInt
   
   private[collection] def newThreshold(_loadFactor: Int, size: Int) = {
     val lf = _loadFactor
